requirements.txt
# Core ML Libraries
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.1.0
imbalanced-learn>=0.12.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.6.0

# Visualization & Reporting
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
joblib>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
